# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CurrentPulse AI - Daily Current Affairs for UPSC/HCS aspirants
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Article {
  id: number;
  title: string;
  source: string;
  /** @nullable */
  sourceUrl?: string | null;
  /** One of: polity, economy, environment, international_relations, science_tech */
  category: string;
  /** GS1, GS2, GS3, GS4 */
  gsMapping: string;
  keyPoints: string[];
  background: string;
  analysis: string;
  /** prelims, mains, both */
  examRelevance: string;
  tags: string[];
  publishedAt: string;
  createdAt: string;
}

export type TodayFeedByCategory = { [key: string]: Article[] };

export interface TodayFeed {
  date: string;
  totalCount: number;
  byCategory: TodayFeedByCategory;
}

export interface Category {
  id: string;
  label: string;
  gsMapping: string;
  articleCount: number;
  description: string;
}

export interface Bookmark {
  id: number;
  articleId: number;
  article: Article;
  createdAt: string;
}

export interface CreateBookmarkBody {
  articleId: number;
}

export interface Note {
  id: number;
  /** @nullable */
  articleId?: number | null;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  /** @nullable */
  articleId?: number | null;
  content: string;
}

export interface UpdateNoteBody {
  content?: string;
}

export interface QuizQuestion {
  id: number;
  articleId: number;
  question: string;
  options: string[];
  /** Index of the correct option (0-3) */
  correctAnswer: number;
  explanation: string;
  /** easy, medium, hard */
  difficulty: string;
  createdAt: string;
}

export type DashboardSummaryCategoryBreakdownItem = {
  category: string;
  count: number;
};

export interface DashboardSummary {
  todayCount: number;
  weekCount: number;
  totalBookmarks: number;
  totalNotes: number;
  categoryBreakdown: DashboardSummaryCategoryBreakdownItem[];
  recentArticles: Article[];
}

export interface TrendingTopic {
  topic: string;
  count: number;
  category: string;
}

export type ListArticlesParams = {
  /**
   * @nullable
   */
  category?: string | null;
  /**
   * ISO date string (YYYY-MM-DD)
   * @nullable
   */
  date?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  examRelevance?: string | null;
};

export type GetWeeklyArticlesParams = {
  /**
   * @nullable
   */
  weekOffset?: number | null;
};

export type ListQuizQuestionsParams = {
  /**
   * @nullable
   */
  articleId?: number | null;
  /**
   * @nullable
   */
  category?: string | null;
};