# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CurrentPulse AI - Daily Current Affairs for UPSC/HCS aspirants
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from "./article";

export interface Bookmark {
  id: number;
  articleId: number;
  article: Article;
  createdAt: Date;
}