# lib/api-zod/src/generated/types/createBookmarkBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CurrentPulse AI - Daily Current Affairs for UPSC/HCS aspirants
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookmarkBody {
  articleId: number;
}