# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CurrentPulse AI - Daily Current Affairs for UPSC/HCS aspirants
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from "./article";
import type { DashboardSummaryCategoryBreakdownItem } from "./dashboardSummaryCategoryBreakdownItem";

export interface DashboardSummary {
  todayCount: number;
  weekCount: number;
  totalBookmarks: number;
  totalNotes: number;
  categoryBreakdown: DashboardSummaryCategoryBreakdownItem[];
  recentArticles: Article[];
}