# lib/api-zod/src/generated/types/todayFeed.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CurrentPulse AI - Daily Current Affairs for UPSC/HCS aspirants
 * OpenAPI spec version: 0.1.0
 */
import type { TodayFeedByCategory } from "./todayFeedByCategory";

export interface TodayFeed {
  date: string;
  totalCount: number;
  byCategory: TodayFeedByCategory;
}